[MERGE #2957 @MikeHolman] Fix bugs from PropertyString cache changes
Merge pull request #2957 from MikeHolman:enumcachefix

Fixes following:
We were not disabling StoreFieldCache for SimpleDictionaryTypeHandle if PropertyDescriptor is not initialized.
We were attempting to cache enumeration of InternalPropertyId in case it was the last property.
We were attempting to cache property reads when object of a LdElem was a tagged number.