[MERGE #2951 @MikeHolman] Add PropertyString pointer to some LiteralStrings
Merge pull request #2951 from MikeHolman:concatpropstr

There are some types of strings which we convert into LiteralString (e.g. ConcatStrings). However, they generally have a number of fields so there is extra space that is unused after conversion. This change utilizes that space with a PropertyString pointer (which we populate if you use the string as an index for a StElem).

This is about a 2% improvement in Speedometer.