[MERGE #2950 @Cellule] IsOpndNegZero bugfix + direct branching
Merge pull request #2950 from Cellule:users/micfer/negzero

There was a bug on x64 for float32 IsNegZero
the conversion to uint32 was not done correctly (MOVQ with uint32 and float32 didn't work).

I took the opportunity to reuse `LowerReinterpretPrimitive` for the conversion and `FindOrCreateInt64Pair` for the x86 double case.
This allowed me to move the code back to LowerMDShared.cpp instead of arch specific implementation between x64 and x86.

Added some test cases to cover the problematic case.
Related #2894