Adding a new diagnostic API to retrieve the console scope object
In the DiagnosticsEngine the intellisense is currently populated with the properties from the the window object. With the console scope there can be variables in that scope that need to be included in the intellisense. This new method returns the console object which can be iterated over for the properties.