[MERGE #3090 @meg-gupta] Avoid adding empty blocks while inserting an edge from finally block to early exit
Merge pull request #3090 from meg-gupta:avoidemptyblocks

When we add an edge from end of finally block to an early exit, we need not create additional blocks
if the Leave/LeaveNull at the end of finally was in a separate block preceeded by a Label.
Creating empty blocks like this can lead to issues with DelateLeaveChains later on, moreover we want to keep the IR clean.

Also piggybacking 2 minor fixes:
1. Don't zero out region info on dead blocks, affects liveness info in the presence of regions
2. Don't call ResetOut on non exception path in ProcessTryHandlerBailout