[MERGE #3303 @dilijev] OS#12416927: Fix line ending of test\Intl\IntlTaintingPreInitTests.baseline to be CRLF
Merge pull request #3303 from dilijev:intl-tainting-crlf

The test baseline was apparently checked-in with LF instead of the intended CRLF, which was fine for most things, which will normalize line endings of files not marked as binary. Nightlies apparently do not do this normalization.

```
$ xxd test/Intl/IntlTaintingPreInitTests.baseline
00000000: 5061 7373 6564 2070 7265 2d69 6e69 7420  Passed pre-init
00000010: 7461 696e 7469 6e67 210d 0a              tainting!..
```

(Note the `0d 0a` at the end, before was only `0a`.)