[MERGE #3244 @Cellule] WASM - x86 int64 gt_s & lt_s
Merge pull request #3244 from Cellule:wasm/x86_int64_cmp

On x86 for int64 compare, we need to compare the low bits using unsigned comparisons.

This removes the need for the hack to prevent cmp src swapping.

Also, lower int64 branching directly instead of relying on cmp then branch on cmp result as it was incorrectly branching on a signed comparison of the lower bits.

Fixes OS#12197588