[MERGE #3333 @Cellule] WASM - int64 split live on back edge fix
Merge pull request #3333 from Cellule:wasm/int64_lifetime

Make sure to mark all int64 split symbol as Live on back edge if the original int64 was also live.

The previous solution to mark them as live when we lower the loop top wasn't working because it is possible for a symbol to be live on back edge, but not seen in lower yet.
```js
for(...) {
  int64 myInt64Var;
  for(...) { // myInt64Var is liveOnBackEdge, but not processed yet (backward pass in lowerer)
    // Need to know that the high and low part of myInt64Var is live on back edge to do compensation in RegAlloc
  }
}
```

Move the int64 split map to Func allowing to use post lower.
Also allowing some lowering methods to be static which was prevented before because it needed an instance of lowerer.

Fixes OS#12708317