[MERGE #3425 @Cellule] Fix possible infinite loop in String.raw()
Merge pull request #3425 from Cellule:raw

I was able to construct a use case where we would end up in an infinite loop.
It takes forever to run to just overflow the uint32, but I figured we should at least avoid possible infinite loops.

A possible solution could be to just cap the length to loop through to UINT32_MAX, but it would be technically not spec compliant.