[MERGE #3493 @MikeHolman] remove inconsistent check in BuildJITTimeData
Merge pull request #3493 from MikeHolman:wasmbody

This can lead to nullptr deref in FunctionJITTimeInfo::BuildJITTimeData in case function is a wasm defer parse function, because of some inconsistency in checks, where one of them was looking for DeferParse attribute to know if function was parsed, which wasm doesn't set.

I looked into setting this, but that causes more issues. Seems safest to just fix the check in BuildJITTimeData. Later, maybe we can change to start as FunctionProxy and convert to FunctionBody after parsing like we do for JS, which will solve the issues that setting the DeferParse flag causes.

OS: 13005931