[MERGE #3587 @rajatd] Setting internal properties with "enumerable" attribute off.
Merge pull request #3587 from rajatd:hasNoEnumerableProps-2

While we don't enumerate internal properties, setting an internal property as enumerable unnecessarily sets the hasNoEnumerableProperties flag to false on the type handler. This may have perf impact as this flag is used to skip prototypes when enumerating properties.

Changing this to set internal properties as non-enumerable from the beginning. Doing this for non-path type handlers only because,
1. An object with a path type handler should theoretically have hasNoEnumerableProperties set to false anyway.
2. Setting a non-enumerable property on an object with a path type handler will convert its type handler to a (simple) dictionary type handler.

Fixes #1622