[MERGE #3608 @Penguinwizzard] Add a flag to use a stronger array sort for bug analysis.
Merge pull request #3608 from Penguinwizzard:strongArraySort

The default JS sort often results in an implementation-defined order.
This is not generally a problem, and helps improve our perf. However,
there are cases where our ordering depends on the internal structure,
not the user-visible structure, of the array. This change adds a flag
to make this ordering more consistent, which should help identify our
bugs where the issue is just our sort orders being inconsistent.