[MERGE #3775 @ricobbe] Implement Dump method on part of DynamicTypeHandler hierarchy
Merge pull request #3775 from ricobbe:build/ricobbe/dictionary-type-handler-base-dump

As part of my current investigation, I need to be able to dump a human-readable represntation of DictionaryTypeHandlerBase objects.  Add a virtual Dump method to DynamicTypeHandle to allow this, and provide stub implementations for other subtypes of DynamicTypeHandler.  (As the need arises, I may provide full implementations for these stubs; others are welcome to do so if they get to it before I do.)