[MERGE #3776 @MikeHolman] fix bug with inlining int64 args on x86
Merge pull request #3776 from MikeHolman:inlineint64arg

In some cases we would change args/ret to be a Ld_A in the inliner. This works in most cases, but wasm wants them to be Ld_I4 for int types and in particular lowering Ld_A of int64 doesn't work.

I've changed to pick a the appropriate opcode depending on type.

OS: 13867303