Move extraction of deviceId up the telemetry chain
This is needed for NGP because we're passing around the device id
in the data; this is a PII-relevant field, so we need it to be an
actual field (as opposed to being buried in a JSON blob) in order
to tag it. We can also throw away the extra data in the parts a/b
of the event; this significantly reduces our NGP-relevant payload
further down the chain.