[PR #3803 @tcare] Remove outdated HTML comment suffix logic from ScriptEngine
Fixes OS: 13801480

A while back we overhauled the scanner logic for HTML comment rules standardized in ES2015. We had one remaining case in ScriptEngine that tries to optimize parsing by removing trailing HTML comments. The rules changed significantly and now depend on other state such as whether we just finished a multiline comment with line breaks.

Simple repro: ///-->

The case in question happened as follows:
- Edge downloads a script that was referenced in a <script src=... tag
- Edge passes that script to us, invoking the CompileUTF8 entrypoint instead of UTF16. I'm not exactly sure why this is, but it always happens when the src attribute was in the script tag.
- The UTF8 entrypoint skips doing a memcpy of the source, so the buffer still contains the HTML comment.
- We incorrectly stop the HTML comment at the first //, leaving just a length of 1 '/'
- The parser starts parsing this as a regex expr with the incorrect length, leading us to trying to try read the following // as a regex.
- The regex parser asserts as we try to read past the incorrect length.

I do not think this is exploitable because you specifically need 3 or more / characters to get the regex parser into this bad state. The regex parser will see the ending / in fre and stop parsing the regex. There is no reading past the buffer because we are guaranteed to have /--> past the regex start.

The fix is to remove the HTML comment optimization and let the scanner handle the complex logic to track whether an HTML comment is acting as a JS comment. I added a test that makes sure we still parse a preceeding regex correctly; note that it requires a UTF8 BOM marker and a <script src= tag to correctly repro.

[Core Commit Message]

[MERGE #3803 @tcare] Clean up unused parser flags

Merge pull request #3803 from tcare:htmlsuffix