[MERGE #3885 @tcare] Fix OS: 13988805 AsmJs error triggers syntax error on reparse assert
Merge pull request #3885 from tcare:asmjsreparse

A recent AsmJs change introduced a throwable AsmJs error at parse time. This error can only be found on a full parse, so if we previously deferred a function that contains this error then we will assert that we should have found all syntax errors in the initial defer parse.

The fix is to special case the AsmJs error, since it is not strictly a syntax error. We will gracefully reparse without AsmJs.