[MERGE #3855 @Penguinwizzard] Remove support for VS2013 (dev12)
Merge pull request #3855 from Penguinwizzard:version_bump

This change removes support for Visual Studio 2013. Doing this will
afford us the use of a much wider set of constructs in C++, as more
recent versions of Visual Studio provide full support for C++11.

I'm going to leave this PR open for a week (at least) to invite any
comments about the change that may be out there.

Note: Although this change is to move to support C++11, we could be
a C++14 code base with little extra effort; our minimum versions of
clang (3.9) and VS (dev14) are both sufficient for C++14 support. I
think the only thing holding that back is that we may want to allow
a 4.9 version of GCC.