[MERGE #2756 @MikeHolman] pass prop string as optional arg for hasOwnProperty; remove dictionary lookup
Merge pull request #2756 from MikeHolman:hasownproparg

HasOwnProperty helper was a bit inefficient where it often gets a PropertyString as an argument, gets the propertyId from that and passes it to another helper which does a dictionary lookup to see if string is cached.

Technically the dictionary lookup might find PropertyString if argument was passed in as some other type of JavascriptString, but mostly we will be there in the negative case, so I removed the dictionary lookup. It's also unclear if that dictionary lookup is even faster than the actual property lookup.

Improves react-redux by 4%.