[MERGE #3941 @thomasmo] Simplify FunctionExecutionStateMachine::TryTransitionToNextExecutionMode
Merge pull request #3941 from thomasmo:infinite_rejit

Simplify FunctionExecutionStateMachine::TryTransitionToNextExecutionMode
This change makes significant changes to FunctionExecutionStateMachine::TryTransitionToNextExecutionMode. The previous implementation of this function was based around the ExecutionMode enum and would jump between states. With the new implementation, a separate enum, ExecutionState, is introduced to have a linear of moving across states. Common code paths across all states are also better shared. With this change, the invariants are consistent across states, and the exceptions stand out more. Further, the API boundary for FunctionExecutionStateMachine and some of FunctionBody is tightened to keep some of the internals of FunctionExecutionStateMachine abstracted. Some of the native unittests in FunctionExecutionTest are refactored to support behavior when FullJit is enabled vs not enabled for better code coverage.