[MERGE #4017 @meg-gupta] Fix clearing callinfo on inlined frames in a try
Merge pull request #4017 from meg-gupta:stackwalktoomanyclears

We inline functions into try, and clear the callinfo of all inlined frames until the enclosing try frame on a throw.
But if the inlined function calls into the interpreter and a throw happens, we do not want to clear the inlinee frame info of outer inlined frames.

In the failing test case, we inlined a function that called into an async function in the interpreter, we incorrectly cleared all the inlined frames until the enclosing try.
And on a following bailout from inside the inlined code in the try, we ended up with cleared callinfo while executing BailOutRecord::BailOutInlinedHelper.

Fixes OS#14226300