Merged PR 91206: Add RemoteRunTests
Add RemoteRunTests

This is a pretty straightforward way to run tests. It loads a Chakra
remote debugging configuration xml file to identify the machine, and
then uses powershell remoting to run tests there. There's a few bits
that complicate it; chief among them is that these machines may have
remote debugging set up, which causes the test runner to lock up; in
order to avoid this, the test runner will continuously kill a windbg
session that starts during the run (and won't run if windbg is going
at the start of the test run).