[MERGE #4062 @xiaoyinl] treat two digit years less than 50 as 21st century years
Merge pull request #4062 from xiaoyinl:2digityear

One year ago, Mozilla changed the way Firefox interprets dates in the format `mm/dd/yy`. In Firefox version 49 or later, if `yy` is less than 50, the year is regarded as `20yy`; if `yy` is 50~99, the year is regarded as `19yy`. The old Firefox behavior matches IE and Edge, and its new behavior matches Chrome and Safari. Given that, I wonder if ChakraCore wants to follow them.

This PR changes the behavior of parsing two digit years to align with SpiderMonkey and v8.

I will update test cases, if you'd like to adopt this behavior change.

/cc @bterlson @dilijev